Combine iterstrings easily

c.iterstring() does exactly what you would expect it to, concatenating two iterstrings and stripping their attributes (except class). insert() inserts an iterstring after pos (remember, R-style indices starting with 1).

Negative indices work the same as with [[.iterstring]—inserting a string at pos of -1 means putting it at the end of the string, i.e. after str[-1], the last character of str.

insert(string, newstring, pos)

# S3 method for iterstring
c(string1, string2)

Arguments

string
the starting iterstring (or character vec of length 1, coerced)

newstring
an iterstring (or character vec. of length 1, coerced) to insert

pos
the position after which newstring will be inserted

string1
the first iterstring in c.iterstring()

string2
the second iterstring in c.iterstring()