R Set Comprehension

Set comprehension with the Magrittr Pipe. Always use the syntax: .x %>% that_for_all(.y) %>% we_have_*(f(.x, .y)).

that_for_all(.x, .y)

that_for_any(.x, .y)

we_have(that_for, formula, result = "vector")

Arguments

.x
A set, represented as either an atomic vector or a list

.y
A set to compare to .x

that_for
A list passed to we_have--can be ignored with proper syntax

formula
A boolean-valued function, lambda, or formula

result
Should the expression return a vector or an Iterator?

Value

For that_for_all and that_for_any, an object of S3 class that_for_all or that_for_any. For we_have, a vector of the same type as .x if return == 'vector' and an Iterator object if return == 'Iterator'.

Details

This function is designed to be used with %>% from magrittr. Essentially, this function evaluates f %>% g() as eval(g(f), envir = new.env(parent = parent.frame()). Any pipe or pipe-like function that works similarly should work with that_for_*. In particular, the R Core Team is expected to implement a base R pipe in the near future with the syntax |>. While these functions will likely work properly with the new |>, as of right now the only pipe that is guaranteed to work is %>%.

Note

if .y is an numeric vector, you probably want a value obtained from peruse::range(start, end) rather than start:end or seq(start,end), as when start is greater than end you want an empty vector rather than counting backwards. Note that peruse::range views end as a supremum, not a maximum, thus range(a,b) is equivalent to the set [a,b) when a < b or {} when b >= a.

Examples

library(magrittr)
2:100 %>% that_for_all(range(2, .x)) %>% we_have(function(.x, .y) .x %% .y != 0) #is the same as

#>  [1]  2  3  5  7 11 13 17 19 23 29 31 37 41 43 47 53 59 61 67 71 73 79 83 89 97

2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0)

#>  [1]  2  3  5  7 11 13 17 19 23 29 31 37 41 43 47 53 59 61 67 71 73 79 83 89 97

#c.f.
primes <- 2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0, "Iterator")
yield_next(primes)

#> [1] 2

{c("I", "Don't", "wan't", "chicken") %>%
             that_for_all("\'") %>%
             we_have(~grepl(.y, .x))}

#> [1] "Don't" "wan't"

#Twin primes 1 through 100
if (FALSE) {
primes <- 2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0)
primes %>% that_for_any(primes) %>% we_have(~abs(.x - .y) == 2)
#Prime numbers 1 through 100 that are two away from a square number
}
(2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0)) %>%
    that_for_any(range(2, .x)) %>% we_have(~sqrt(.x + 2) == .y | sqrt(.x - 2) == .y)

#> [1]  7 11 23 47 79 83