Data with hash table internals

Create a data.frame built on the environment S3 class and embedded in an R6 object. The primary advantage of this structure is the reference semantics, which means that elements (in this case data.frame columns) are not copied upon modification. Although modern R avoids making deep copies of data.frames when they are modified, the copy on modify semantics applied to the container (the list of pointers to the columns of the data.frame), can still significantly slow down loops that modify large numbers of columns on wide data.frames.

hash_dfs also offer access to their columns in amortized constant time (O(1)) due to their internal structure as a hash table.

hash_dfs are bare-bones, with methods for adding (binding) and deleting (unbinding) variables as well as printing a preview of the data. In addition, dplyr functions select and mutate (as well as its scoped versions) have implementations for hash_df's (that do not depend on {dplyr}). Note that in {dplyr}, across() and tidyselect() have superseded scoped verbs. However, hash_dfs do not understand tidyselect, so we must rely on the scoped versions (mutate_*) when wrangling these objects. Data wrangling and analysis functions not implemented in hash_df should be done after converting back to a data.frame object with hash_df$return_df().

Value

A hash_df object, built on R6 and env S3 classes

See also

Public fields

data

an environment containing the the data from a data.frame.

Active bindings

nrow

the number of rows of the data.frame

ncol

the number of columns of the data.frame

vars

the variables (column names) of the data.frame

Methods

Public methods

hash_df$bind()

hash_df$unbind()

hash_df$return_df()

hash_df$print()

hash_df$new()

hash_df$View()

hash_df$mutate()

hash_df$mutate_if()

hash_df$mutate_all()

hash_df$mutate_at()

hash_df$select()

hash_df$select_if()

hash_df$select_at()

hash_df$clone()

Method bind()

add data.frame columns

Usage

hash_df$bind(...)

Arguments

...

named columns to add

Examples

df <- hash_df$new(iris)
df$bind(col_ones = 1, col_his = "hi")
df$print()

Method unbind()

remove data.frame columns

Usage

hash_df$unbind(...)

Arguments

...

names of the columns to remove

Examples

df <- hash_df$new(iris)
df$unbind(Petal.Length)
df$print()

Method return_df()

return the data.frame from a hash_df

Usage

hash_df$return_df()

Returns

A data.frame object

Method print()

print a preview of the data.frame

Usage

hash_df$print()

Method new()

create a new hash_df from a data.frame

Usage

hash_df$new(df)

Arguments

df

a data.frame

Method View()

open the data in the View pane

Usage

hash_df$View()

Method mutate()

a bare-bones dplyr::mutate that takes advantage of the hashed data structure

Usage

hash_df$mutate(...)

Arguments

...

named R expressions to mutate the columns by

Examples

df <- hash_df$new(iris)
df$mutate(Sepal.Size = Sepal.Length * Sepal.Width)
df$print()

Method mutate_if()

Think dplyr::mutate_if or mutate(across(where(...))), but only taking one function, not a list. For all the scoped mutates in {dplyr}, functions can be functions/lambdas, formulas, or certain syntactic sugar expressions like mean --> ~mean(.x). See purrr::as_mapper for more information about function conversions.

Usage

hash_df$mutate_if(test, fun)

Arguments

test

a predicate function

fun

a function to apply when test is TRUE

Examples

df <- hash_df$new(iris)
df$mutate_if(is.numeric, log)
df$print()

Method mutate_all()

Think dplyr::mutate_all or mutate(across(everything, ...)), but again only taking one function, not a list.

Usage

hash_df$mutate_all(fun)

Arguments

fun

a function to apply to every variable

Examples

df <- hash_df$new(mtcars)
df$mutate_all(~.x - mean(.x))
df$print()

Method mutate_at()

Think dplyr::mutate_at, but again only taking one function, not a list. Also does not understand tidyselect syntax, so regex must be a regular expression that is understood by base::grepl.

Usage

hash_df$mutate_at(regex, fun, ...)

Arguments

regex

a regular expression

fun

a function to apply to variables whose names match regex

...

additional arguments passed to regex

Examples

df <- hash_df$new(iris)
df$mutate_at("Sepal*", ~.x / 100)
df$print()

Method select()

A bare-bones dplyr::select that takes advantage of the hashed data structure

Usage

hash_df$select(...)

Arguments

...

variables to select, separated by a comma

Examples

df <- hash_df$new(iris)
df$select(Sepal.Length, Species)
df$print()

Method select_if()

Think dplyr::select_if. Note that peruse::hash_df$select_if() does not take advantage of the particular hashed data type and is only included for completeness. There is no reason to think that this will be more efficient than dplyr::select_if, and it may be less efficient.

Usage

hash_df$select_if(test)

Arguments

test

a predicate function to subset data

Examples

df <- hash_df$new(iris)
df$select_if(is.numeric)
df$print()

Method select_at()

Think dplyr::select_at, but again without tidyselect specification, so regex must be understood by grepl.

Usage

hash_df$select_at(regex)

Arguments

regex

a regular expression to subset the data

Examples

df <- hash_df$new(iris)
df$select_at("Sepal*")
df$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

hash_df$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

df <- hash_df$new(iris)
for (name in df$vars) {
 if (is.numeric(df$data[[name]])) df$data[[name]] <- 2 * df$data[[name]]
}

hash_mtcars <- hash_df$new(mtcars)
hash_mtcars$data <- lapply(hash_mtcars$data, function(x) log(x + 1))
hash_mtcars$print()

#>         carb     gear        am        vs     qsec        wt     drat       hp
#> 1  1.6094379 1.609438 0.6931472 0.0000000 2.859913 1.2864740 1.589235 4.709530
#> 2  1.6094379 1.609438 0.6931472 0.0000000 2.891482 1.3545457 1.589235 4.709530
#> 3  0.6931472 1.609438 0.6931472 0.6931472 2.976040 1.1999648 1.578979 4.543295
#> 4  0.6931472 1.386294 0.0000000 0.6931472 3.017494 1.4386496 1.406097 4.709530
#> 5  1.0986123 1.386294 0.0000000 0.0000000 2.891482 1.4906544 1.423108 5.170484
#> 6  0.6931472 1.386294 0.0000000 0.6931472 3.054944 1.4951488 1.324419 4.663439
#> 7  1.6094379 1.386294 0.0000000 0.0000000 2.823757 1.5195132 1.437463 5.505332
#> 8  1.0986123 1.609438 0.0000000 0.6931472 3.044522 1.4327007 1.545433 4.143135
#> 9  1.0986123 1.609438 0.0000000 0.6931472 3.173878 1.4231083 1.593309 4.564348
#> 10 1.6094379 1.609438 0.0000000 0.6931472 2.960105 1.4906544 1.593309 4.820282
#> 11 1.6094379 1.609438 0.0000000 0.6931472 2.990720 1.4906544 1.593309 4.820282
#> 12 1.3862944 1.386294 0.0000000 0.0000000 2.912351 1.6233408 1.403643 5.198497
#> 13 1.3862944 1.386294 0.0000000 0.0000000 2.923162 1.5539252 1.403643 5.198497
#> 14 1.3862944 1.386294 0.0000000 0.0000000 2.944439 1.5644405 1.403643 5.198497
#> 15 1.6094379 1.386294 0.0000000 0.0000000 2.943386 1.8325815 1.368639 5.327876
#> 16 1.6094379 1.386294 0.0000000 0.0000000 2.934920 1.8600410 1.386294 5.375278
#> 17 1.6094379 1.386294 0.0000000 0.0000000 2.913437 1.8476671 1.442202 5.442418
#> 18 0.6931472 1.609438 0.6931472 0.6931472 3.018960 1.1631508 1.625311 4.204693
#> 19 1.0986123 1.609438 0.6931472 0.6931472 2.971440 0.9612641 1.780024 3.970292
#> 20 0.6931472 1.609438 0.6931472 0.6931472 3.039749 1.0420419 1.652497 4.189655
#>        disp      cyl      mpg
#> 1  5.081404 1.945910 3.091042
#> 2  5.081404 1.945910 3.091042
#> 3  4.691348 1.609438 3.169686
#> 4  5.556828 1.945910 3.109061
#> 5  5.888878 2.197225 2.980619
#> 6  5.420535 1.945910 2.949688
#> 7  5.888878 2.197225 2.727853
#> 8  4.995183 1.609438 3.234749
#> 9  4.954418 1.609438 3.169686
#> 10 5.127529 1.945910 3.005683
#> 11 5.127529 1.945910 2.933857
#> 12 5.623295 2.197225 2.856470
#> 13 5.623295 2.197225 2.906901
#> 14 5.623295 2.197225 2.785011
#> 15 6.159095 2.197225 2.433613
#> 16 6.133398 2.197225 2.433613
#> 17 6.089045 2.197225 2.753661
#> 18 4.378270 1.609438 3.508556
#> 19 4.339902 1.609438 3.446808
#> 20 4.278054 1.609438 3.552487

wide_df <- as.data.frame(matrix(1:5000, nrow = 2))
hash_wide_df <- hash_df$new(wide_df)

# using a `for` loop:
for (name in hash_wide_df$vars) {
  hash_wide_df$data[[name]] <- -1 *  hash_wide_df$data[[name]]
}

# using `lapply`:
hash_wide_df$data <- lapply(hash_wide_df$data, function(x) x * -1)

# using `mutate`:
hash_wide_df$mutate(~.x * -1)

if (FALSE) {
timer <- bench::mark(
 check = F,
 for(name in names(wide_df)) wide_df[[name]] <- -1 * wide_df[[name]],
 for (name in hash_wide_df$vars) hash_wide_df$data[[name]] <- -1 *  hash_wide_df$data[[name]]
)}

#> hash_df: Median = 22.2.3ms
#> data.frame: Median = 615.3ms

## ------------------------------------------------
## Method `hash_df$bind`
## ------------------------------------------------

df <- hash_df$new(iris)
df$bind(col_ones = 1, col_his = "hi")
df$print()

#>    col_his col_ones Species Petal.Width Petal.Length Sepal.Width Sepal.Length
#> 1       hi        1  setosa         0.2          1.4         3.5          5.1
#> 2       hi        1  setosa         0.2          1.4         3.0          4.9
#> 3       hi        1  setosa         0.2          1.3         3.2          4.7
#> 4       hi        1  setosa         0.2          1.5         3.1          4.6
#> 5       hi        1  setosa         0.2          1.4         3.6          5.0
#> 6       hi        1  setosa         0.4          1.7         3.9          5.4
#> 7       hi        1  setosa         0.3          1.4         3.4          4.6
#> 8       hi        1  setosa         0.2          1.5         3.4          5.0
#> 9       hi        1  setosa         0.2          1.4         2.9          4.4
#> 10      hi        1  setosa         0.1          1.5         3.1          4.9
#> 11      hi        1  setosa         0.2          1.5         3.7          5.4
#> 12      hi        1  setosa         0.2          1.6         3.4          4.8
#> 13      hi        1  setosa         0.1          1.4         3.0          4.8
#> 14      hi        1  setosa         0.1          1.1         3.0          4.3
#> 15      hi        1  setosa         0.2          1.2         4.0          5.8
#> 16      hi        1  setosa         0.4          1.5         4.4          5.7
#> 17      hi        1  setosa         0.4          1.3         3.9          5.4
#> 18      hi        1  setosa         0.3          1.4         3.5          5.1
#> 19      hi        1  setosa         0.3          1.7         3.8          5.7
#> 20      hi        1  setosa         0.3          1.5         3.8          5.1

## ------------------------------------------------
## Method `hash_df$unbind`
## ------------------------------------------------

df <- hash_df$new(iris)
df$unbind(Petal.Length)
df$print()

#>    Species Petal.Width Sepal.Width Sepal.Length
#> 1   setosa         0.2         3.5          5.1
#> 2   setosa         0.2         3.0          4.9
#> 3   setosa         0.2         3.2          4.7
#> 4   setosa         0.2         3.1          4.6
#> 5   setosa         0.2         3.6          5.0
#> 6   setosa         0.4         3.9          5.4
#> 7   setosa         0.3         3.4          4.6
#> 8   setosa         0.2         3.4          5.0
#> 9   setosa         0.2         2.9          4.4
#> 10  setosa         0.1         3.1          4.9
#> 11  setosa         0.2         3.7          5.4
#> 12  setosa         0.2         3.4          4.8
#> 13  setosa         0.1         3.0          4.8
#> 14  setosa         0.1         3.0          4.3
#> 15  setosa         0.2         4.0          5.8
#> 16  setosa         0.4         4.4          5.7
#> 17  setosa         0.4         3.9          5.4
#> 18  setosa         0.3         3.5          5.1
#> 19  setosa         0.3         3.8          5.7
#> 20  setosa         0.3         3.8          5.1

## ------------------------------------------------
## Method `hash_df$mutate`
## ------------------------------------------------

df <- hash_df$new(iris)
df$mutate(Sepal.Size = Sepal.Length * Sepal.Width)
df$print()

#>    Sepal.Size Species Petal.Width Petal.Length Sepal.Width Sepal.Length
#> 1       17.85  setosa         0.2          1.4         3.5          5.1
#> 2       14.70  setosa         0.2          1.4         3.0          4.9
#> 3       15.04  setosa         0.2          1.3         3.2          4.7
#> 4       14.26  setosa         0.2          1.5         3.1          4.6
#> 5       18.00  setosa         0.2          1.4         3.6          5.0
#> 6       21.06  setosa         0.4          1.7         3.9          5.4
#> 7       15.64  setosa         0.3          1.4         3.4          4.6
#> 8       17.00  setosa         0.2          1.5         3.4          5.0
#> 9       12.76  setosa         0.2          1.4         2.9          4.4
#> 10      15.19  setosa         0.1          1.5         3.1          4.9
#> 11      19.98  setosa         0.2          1.5         3.7          5.4
#> 12      16.32  setosa         0.2          1.6         3.4          4.8
#> 13      14.40  setosa         0.1          1.4         3.0          4.8
#> 14      12.90  setosa         0.1          1.1         3.0          4.3
#> 15      23.20  setosa         0.2          1.2         4.0          5.8
#> 16      25.08  setosa         0.4          1.5         4.4          5.7
#> 17      21.06  setosa         0.4          1.3         3.9          5.4
#> 18      17.85  setosa         0.3          1.4         3.5          5.1
#> 19      21.66  setosa         0.3          1.7         3.8          5.7
#> 20      19.38  setosa         0.3          1.5         3.8          5.1

## ------------------------------------------------
## Method `hash_df$mutate_if`
## ------------------------------------------------

df <- hash_df$new(iris)
df$mutate_if(is.numeric, log)
df$print()

#>    Species Petal.Width Petal.Length Sepal.Width Sepal.Length
#> 1   setosa  -1.6094379   0.33647224    1.252763     1.629241
#> 2   setosa  -1.6094379   0.33647224    1.098612     1.589235
#> 3   setosa  -1.6094379   0.26236426    1.163151     1.547563
#> 4   setosa  -1.6094379   0.40546511    1.131402     1.526056
#> 5   setosa  -1.6094379   0.33647224    1.280934     1.609438
#> 6   setosa  -0.9162907   0.53062825    1.360977     1.686399
#> 7   setosa  -1.2039728   0.33647224    1.223775     1.526056
#> 8   setosa  -1.6094379   0.40546511    1.223775     1.609438
#> 9   setosa  -1.6094379   0.33647224    1.064711     1.481605
#> 10  setosa  -2.3025851   0.40546511    1.131402     1.589235
#> 11  setosa  -1.6094379   0.40546511    1.308333     1.686399
#> 12  setosa  -1.6094379   0.47000363    1.223775     1.568616
#> 13  setosa  -2.3025851   0.33647224    1.098612     1.568616
#> 14  setosa  -2.3025851   0.09531018    1.098612     1.458615
#> 15  setosa  -1.6094379   0.18232156    1.386294     1.757858
#> 16  setosa  -0.9162907   0.40546511    1.481605     1.740466
#> 17  setosa  -0.9162907   0.26236426    1.360977     1.686399
#> 18  setosa  -1.2039728   0.33647224    1.252763     1.629241
#> 19  setosa  -1.2039728   0.53062825    1.335001     1.740466
#> 20  setosa  -1.2039728   0.40546511    1.335001     1.629241

## ------------------------------------------------
## Method `hash_df$mutate_all`
## ------------------------------------------------

df <- hash_df$new(mtcars)
df$mutate_all(~.x - mean(.x))
df$print()

#>       carb    gear       am      vs     qsec       wt       drat       hp
#> 1   1.1875  0.3125  0.59375 -0.4375 -1.38875 -0.59725  0.3034375 -36.6875
#> 2   1.1875  0.3125  0.59375 -0.4375 -0.82875 -0.34225  0.3034375 -36.6875
#> 3  -1.8125  0.3125  0.59375  0.5625  0.76125 -0.89725  0.2534375 -53.6875
#> 4  -1.8125 -0.6875 -0.40625  0.5625  1.59125 -0.00225 -0.5165625 -36.6875
#> 5  -0.8125 -0.6875 -0.40625 -0.4375 -0.82875  0.22275 -0.4465625  28.3125
#> 6  -1.8125 -0.6875 -0.40625  0.5625  2.37125  0.24275 -0.8365625 -41.6875
#> 7   1.1875 -0.6875 -0.40625 -0.4375 -2.00875  0.35275 -0.3865625  98.3125
#> 8  -0.8125  0.3125 -0.40625  0.5625  2.15125 -0.02725  0.0934375 -84.6875
#> 9  -0.8125  0.3125 -0.40625  0.5625  5.05125 -0.06725  0.3234375 -51.6875
#> 10  1.1875  0.3125 -0.40625  0.5625  0.45125  0.22275  0.3234375 -23.6875
#> 11  1.1875  0.3125 -0.40625  0.5625  1.05125  0.22275  0.3234375 -23.6875
#> 12  0.1875 -0.6875 -0.40625 -0.4375 -0.44875  0.85275 -0.5265625  33.3125
#> 13  0.1875 -0.6875 -0.40625 -0.4375 -0.24875  0.51275 -0.5265625  33.3125
#> 14  0.1875 -0.6875 -0.40625 -0.4375  0.15125  0.56275 -0.5265625  33.3125
#> 15  1.1875 -0.6875 -0.40625 -0.4375  0.13125  2.03275 -0.6665625  58.3125
#> 16  1.1875 -0.6875 -0.40625 -0.4375 -0.02875  2.20675 -0.5965625  68.3125
#> 17  1.1875 -0.6875 -0.40625 -0.4375 -0.42875  2.12775 -0.3665625  83.3125
#> 18 -1.8125  0.3125  0.59375  0.5625  1.62125 -1.01725  0.4834375 -80.6875
#> 19 -0.8125  0.3125  0.59375  0.5625  0.67125 -1.60225  1.3334375 -94.6875
#> 20 -1.8125  0.3125  0.59375  0.5625  2.05125 -1.38225  0.6234375 -81.6875
#>           disp     cyl       mpg
#> 1   -70.721875 -0.1875  0.909375
#> 2   -70.721875 -0.1875  0.909375
#> 3  -122.721875 -2.1875  2.709375
#> 4    27.278125 -0.1875  1.309375
#> 5   129.278125  1.8125 -1.390625
#> 6    -5.721875 -0.1875 -1.990625
#> 7   129.278125  1.8125 -5.790625
#> 8   -84.021875 -2.1875  4.309375
#> 9   -89.921875 -2.1875  2.709375
#> 10  -63.121875 -0.1875 -0.890625
#> 11  -63.121875 -0.1875 -2.290625
#> 12   45.078125  1.8125 -3.690625
#> 13   45.078125  1.8125 -2.790625
#> 14   45.078125  1.8125 -4.890625
#> 15  241.278125  1.8125 -9.690625
#> 16  229.278125  1.8125 -9.690625
#> 17  209.278125  1.8125 -5.390625
#> 18 -152.021875 -2.1875 12.309375
#> 19 -155.021875 -2.1875 10.309375
#> 20 -159.621875 -2.1875 13.809375

## ------------------------------------------------
## Method `hash_df$mutate_at`
## ------------------------------------------------

df <- hash_df$new(iris)
df$mutate_at("Sepal*", ~.x / 100)
df$print()

#>    Species Petal.Width Petal.Length Sepal.Width Sepal.Length
#> 1   setosa         0.2          1.4       0.035        0.051
#> 2   setosa         0.2          1.4       0.030        0.049
#> 3   setosa         0.2          1.3       0.032        0.047
#> 4   setosa         0.2          1.5       0.031        0.046
#> 5   setosa         0.2          1.4       0.036        0.050
#> 6   setosa         0.4          1.7       0.039        0.054
#> 7   setosa         0.3          1.4       0.034        0.046
#> 8   setosa         0.2          1.5       0.034        0.050
#> 9   setosa         0.2          1.4       0.029        0.044
#> 10  setosa         0.1          1.5       0.031        0.049
#> 11  setosa         0.2          1.5       0.037        0.054
#> 12  setosa         0.2          1.6       0.034        0.048
#> 13  setosa         0.1          1.4       0.030        0.048
#> 14  setosa         0.1          1.1       0.030        0.043
#> 15  setosa         0.2          1.2       0.040        0.058
#> 16  setosa         0.4          1.5       0.044        0.057
#> 17  setosa         0.4          1.3       0.039        0.054
#> 18  setosa         0.3          1.4       0.035        0.051
#> 19  setosa         0.3          1.7       0.038        0.057
#> 20  setosa         0.3          1.5       0.038        0.051

## ------------------------------------------------
## Method `hash_df$select`
## ------------------------------------------------

df <- hash_df$new(iris)
df$select(Sepal.Length, Species)
df$print()

#>    Species Sepal.Length
#> 1   setosa          5.1
#> 2   setosa          4.9
#> 3   setosa          4.7
#> 4   setosa          4.6
#> 5   setosa          5.0
#> 6   setosa          5.4
#> 7   setosa          4.6
#> 8   setosa          5.0
#> 9   setosa          4.4
#> 10  setosa          4.9
#> 11  setosa          5.4
#> 12  setosa          4.8
#> 13  setosa          4.8
#> 14  setosa          4.3
#> 15  setosa          5.8
#> 16  setosa          5.7
#> 17  setosa          5.4
#> 18  setosa          5.1
#> 19  setosa          5.7
#> 20  setosa          5.1

## ------------------------------------------------
## Method `hash_df$select_if`
## ------------------------------------------------

df <- hash_df$new(iris)
df$select_if(is.numeric)
df$print()

#>    Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1           5.1         3.5          1.4         0.2
#> 2           4.9         3.0          1.4         0.2
#> 3           4.7         3.2          1.3         0.2
#> 4           4.6         3.1          1.5         0.2
#> 5           5.0         3.6          1.4         0.2
#> 6           5.4         3.9          1.7         0.4
#> 7           4.6         3.4          1.4         0.3
#> 8           5.0         3.4          1.5         0.2
#> 9           4.4         2.9          1.4         0.2
#> 10          4.9         3.1          1.5         0.1
#> 11          5.4         3.7          1.5         0.2
#> 12          4.8         3.4          1.6         0.2
#> 13          4.8         3.0          1.4         0.1
#> 14          4.3         3.0          1.1         0.1
#> 15          5.8         4.0          1.2         0.2
#> 16          5.7         4.4          1.5         0.4
#> 17          5.4         3.9          1.3         0.4
#> 18          5.1         3.5          1.4         0.3
#> 19          5.7         3.8          1.7         0.3
#> 20          5.1         3.8          1.5         0.3

## ------------------------------------------------
## Method `hash_df$select_at`
## ------------------------------------------------

df <- hash_df$new(iris)
df$select_at("Sepal*")
df$print()

#>    Sepal.Width Sepal.Length
#> 1          3.5          5.1
#> 2          3.0          4.9
#> 3          3.2          4.7
#> 4          3.1          4.6
#> 5          3.6          5.0
#> 6          3.9          5.4
#> 7          3.4          4.6
#> 8          3.4          5.0
#> 9          2.9          4.4
#> 10         3.1          4.9
#> 11         3.7          5.4
#> 12         3.4          4.8
#> 13         3.0          4.8
#> 14         3.0          4.3
#> 15         4.0          5.8
#> 16         4.4          5.7
#> 17         3.9          5.4
#> 18         3.5          5.1
#> 19         3.8          5.7
#> 20         3.8          5.1