Functions to create and manipulate islists, lists of iterstrings

Since an iterstring is an S3 object, there can be no atomic vectors of iterstrings. Therefore containers of iterstrings are lists, and their S3 class is islist. An islist is just a list of iterstrings, so it is inadvisable to try to call as_islist on any object except a character vector, a list of iterstrings, or a list of character vectors each of length one.

# S3 method for islist
as.character(islist)

as_islist(object)

Arguments

islist
an islist to turn into a character vector

object
an object to convert to islist